# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DevOps Secrets Manager API
  version: 1.0.0
  description: |
    Manage encrypted secrets organised into vaults and environments, with per-vault roles and
    an audit log.

    Field names are snake_case. Timestamps are RFC 3339. Every error response has the
    `Error` shape. The web app reaches this API under `/api` (proxied by Vite or nginx); the
    API itself has no prefix.

    Callers with no role on a vault get `404` for the vault and everything in it, so IDs that
    belong to other teams are never confirmed. Callers who can see a vault but whose role lacks
    a permission get `403`.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development API
  - url: http://localhost:5173/api
    description: Local web app proxy
security:
  - bearerAuth: []
...
paths:
...
  /envs/{id}:
    parameters:
      - $ref: '#/components/parameters/EnvironmentIdPath'
    get:
      tags: [Environments]
      summary: Get environment
      description: Requires read permission on the environment's vault.
      operationId: getEnvironment
      responses:
        '500': { $ref: '#/components/responses/InternalError' }
        '200':
          description: Environment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Environment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Environments]
      summary: Update environment
      description: Requires write permission.
      operationId: updateEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnvironmentRequest' }
      responses:
        '500': { $ref: '#/components/responses/InternalError' }
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Environment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Environments]
      summary: Delete environment
      description: Requires write permission. The environment is soft-deleted.
      operationId: deleteEnvironment
      responses:
        '500': { $ref: '#/components/responses/InternalError' }
        '204':
          description: Deleted
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /envs/{id}/tokens:
    parameters:
      - $ref: '#/components/parameters/EnvironmentIdPath'
    get:
      tags: [Service tokens]
      summary: List service tokens
      description: Active tokens for the environment (never their values). Requires manage-members permission on the vault.
      operationId: listServiceTokens
      responses:
        '500': { $ref: '#/components/responses/InternalError' }
        '200':
          description: Tokens
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ServiceToken' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Service tokens]
      summary: Create a service token
      description: Issues a read-only token for this environment. The `token` value is returned only in this response; the API stores a SHA-256 hash. Requires manage-members permission on the vault.
      operationId: createServiceToken
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateServiceTokenRequest' }
      responses:
        '500': { $ref: '#/components/responses/InternalError' }
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedServiceToken' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tokens/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Service tokens]
      summary: Revoke a service token
      operationId: revokeServiceToken
      responses:
        '500': { $ref: '#/components/responses/InternalError' }
        '204':
          description: Revoked
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      description: A service token (dsm_st_...) created for one environment.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 access token from /auth/login or /auth/refresh (15 minutes by default).
  parameters:
...
    EnvironmentIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    InternalError:
      description: Unexpected server error (`internal_error`); details are logged, never returned
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Invalid ID, query parameter or body (`invalid_request`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid or expired access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The caller's role does not allow this action (`forbidden`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found, or the caller has no access to it (`not_found`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: A resource with that name already exists (`duplicate` or `conflict`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable code, e.g. not_found, forbidden, invalid_request
          example: not_found
        message:
          type: string
          example: Vault not found
...
    Environment:
      type: object
      required: [id, vault_id, name, description, created_at, updated_at, secret_count]
      properties:
        id: { type: string, format: uuid }
        vault_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        secret_count:
          type: integer
          description: Live secrets; only filled in by the list endpoint
    EnvironmentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
          example: staging
        description: { type: string, nullable: true }
...
    ServiceToken:
      type: object
      required: [id, name, prefix, environment_id, environment_name, vault_id, vault_name, created_by, created_at, expires_at, last_used_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        prefix:
          type: string
          description: The first characters of the token, to recognise it
        environment_id: { type: string, format: uuid }
        environment_name: { type: string }
        vault_id: { type: string, format: uuid }
        vault_name: { type: string }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, nullable: true }
        last_used_at: { type: string, format: date-time, nullable: true }
    CreatedServiceToken:
      type: object
      required: [id, name, prefix, environment_id, environment_name, vault_id, vault_name, created_by, created_at, expires_at, last_used_at, token]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        prefix: { type: string }
        environment_id: { type: string, format: uuid }
        environment_name: { type: string }
        vault_id: { type: string, format: uuid }
        vault_name: { type: string }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, nullable: true }
        last_used_at: { type: string, format: date-time, nullable: true }
        token:
          type: string
          description: The token itself. Shown only once.
    CreateServiceTokenRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        expires_in_days: { type: integer, minimum: 1, maximum: 365, nullable: true }
...